+++ src/components/TypeInput/TypeInput.tsx
@@ -3,9 +3,9 @@ import { InputProps, Select } from "@mantine/core";
 
 const types = ["movie", "series", "episode"];
 
-type Props = {
+interface Props {
   type: InputProps;
-};
+}
 
 const TypeInput = ({ type }: Props) => {
   return <Select allowDeselect placeholder="Type" data={types} {...type} />;
