+++ src/components/YearInput/YearInput.tsx
@@ -1,9 +1,9 @@
 import React from "react";
 import { TextInput, InputProps } from "@mantine/core";
 
-type Props = {
+interface Props {
   year: InputProps;
-};
+}
 
 const YearInput = ({ year }: Props) => {
   return <TextInput type="number" placeholder="Release year" {...year} />;
