--- src/interfaces/Card.ts
@@ -1,7 +1,0 @@
-export interface Card {
-  poster: string;
-  title: string;
-  type: string;
-  year: string;
-  imdbID: string;
-}
