--- src/interfaces/Search.ts
@@ -1,11 +1,0 @@
-export interface SearchQueryParams {
-  search: string;
-  year: string;
-  type: string;
-}
-
-export interface SearchFormValues {
-  search: string;
-  year: string;
-  type: string;
-}
