--- src/interfaces/index.ts
@@ -1,2 +1,0 @@
-export * from "./Card";
-export * from "./Search";
